requirements.txt
python-telegram-bot==20.3
yt-dlp==2023.09.24
